refactor(mpsc): keep bounded receiver waiting local to the ring
